Move commandHandler to zoneserver
